requirements.txt
# Core
Django>=5.0,<6.0
psycopg2-binary>=2.9
Pillow>=10.0

# Kafka
kafka-python>=2.0

# Production
gunicorn>=21.2
whitenoise>=6.5
